+++ src/GameObjects/Enemies/MortalEnemy.cpp
@@ -15,16 +15,16 @@ game::MortalEnemy::MortalEnemy(int posX, int posY) {
 void game::MortalEnemy::move() {
     switch (direction) {
         case moveUp:
-            this->pos.y -= (float) moveSpeed;
+            this->pos.y -= moveSpeed;
             break;
         case moveLeft:
-            this->pos.x -= (float) moveSpeed;
+            this->pos.x -= moveSpeed;
             break;
         case moveDown:
-            this->pos.y += (float) moveSpeed;
+            this->pos.y += moveSpeed;
             break;
         case moveRight:
-            this->pos.x += (float) moveSpeed;
+            this->pos.x += moveSpeed;
             break;
     }
 }
