+++ src/GameObjects/Memories.cpp
@@ -7,6 +7,6 @@
 
 
 Game::Memory::Memory(int memoryX, int memoryY) {
-    setPos(memoryX, memoryY);
+    setPos((float) memoryX, (float) memoryY);
     setCollRec(this->getPos().x, this->getPos().y, 24, 24);
 }
