+++ src/GameObjects/Riegel.cpp
@@ -1,9 +1,8 @@
-#include <iostream>
 #include "Riegel.h"
 
 
 Game::Riegel::Riegel(int RiegelX, int RiegelY) {
-    this->setPos(RiegelX, RiegelY);
+    this->setPos((float) RiegelX, (float) RiegelY);
 
 }
 
@@ -30,7 +29,7 @@ void Game::Riegel::move() {
     if (riegelCanMove) {
         if (CheckCollisionPointRec(mousePosition, {pos.x, pos.y, 24, 24}) &&
             IsMouseButtonDown(MOUSE_LEFT_BUTTON)) {
-            this->pos.x = mousePosition.x - 24 / 2;
+            this->pos.x = mousePosition.x - (float) 24 / 2;
         }
     }
 }
